Adds new EFFECT_TYPE_REVOKE_LAW
Adds new EFFECT_TYPE_REVOKE_LAW

New effect type EFFECT_TYPE_REVOKE_LAW has been added to rules engine.
When applied, the effect of the law is to remove the law that was triggered.
Can be used to create quests that reward a single player for a completing a specified event.