Adds game message to players when new law has been submitted
Adds game message to players when new law has been submitted

Adds notification to players when a new law has been submitted. Under current system,
when laws are submitted, they can be passed with a single aye.  With introduction of the notification, it gives players a chance to review new laws and vote against the ones they do not agree with.